be/requirements.txt
# Wishing Well Backend Requirements

# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Machine Learning
bertopic>=0.16.0
sentence-transformers>=2.2.0
umap-learn>=0.5.0
hdbscan>=0.8.0
scikit-learn>=1.3.0

# NLP
nltk>=3.8.0
gensim>=4.3.0

# OpenAI
openai>=1.0.0

# Scheduler
apscheduler>=3.10.0

# Configuration
python-dotenv>=1.0.0

# Visualization (for testing)
pyldavis>=3.3.0
